Added Rspec to DBC Skeleton
In order to do testing in the sinatra skeleton, we need to have rspec added. There are several things that needto happen.

1. Create a spec folder
2. Create a spec_helper. (This was copied from another project. I forget which one. )
   Specifcally, these magic words needed to go in the spec_helper.rb file

  require 'rubygems'

  # All our specs should require 'spec_helper' (this file)

  # If RACK_ENV isn't set, set it to 'test'.  Sinatra defaults to development,
  # so we have to override that unless we want to set RACK_ENV=test from the
  # command line when we run rake spec.  That's tedious, so do it here.
  ENV['RACK_ENV'] ||= 'test'

  require File.expand_path("../../config/environment", __FILE__)
3. create a migration using rake generate migration task
4. create a model using rake generate model task
5. create a spec for your model using rake generate spec task
6. If you try to create a new model in your spec file, I encourage you to do so without setting up your database. You will get an entertaining series of errors that are fun to squash using error driven development.  However, one in particular was difficult to solve:
     PG::Error:
       FATAL:  database "Sinatra_Basic_rspec_test" does not exist

Our rspec file was expecting a 'test' database, while our databse was named Sinatra_Basic_rspec_development.  Supposidly, the spec_helper fixes that issue by setting a 'test' flag, which implies that a test database should be created.  However, it wasnt. So whe set ENV['RACK_ENV'] ||= 'test' on line five of the environment.rb file.

This is a terrible hack, and I expect that the next commit will be all about how to properly switch between a testing environment and a development environment.

Still, we have kludged tests into RSpec, and are therefore very proud our ourselves.

Tara & Ben