Created test of bcrypt hashes.
Note: the reload function forces the object to reload info from the database.
If you don't have the reload function, this test passes, but it passes because
a bcrypt object does not equal a string, NOT because a hashed string does not equal
a string.

It's like saying that a truck does not equal a pen when what you wanted to
say was that a pencil does not equal a pen.