Usernames are now unique
Created a unique validation in 'rubyland.'

An activerecord validation is like a doorman, and the database is like a building.  The doorman makes sure to throw out anybody that doesn't belong before they get past the door.

An activerecord table constraint is like everybody in the building looking at
a stranger and knowing they don't belong, so they get together and throw
the stranger out.

The first is less secure than the latter.  The latter is really hard to change later. Changing a database constraint is like trying to change a community's dislike of strangers, while changing an activerecord validation is as easy as hiring a new doorman.