How do I write a basic where method?
A where method is like a dude at a warehouse who's job it is to go and find things in that warehouse.  He usually gets a vague descirption, like 'the blue box' or 'every box that is blue.'  He brings a bucket full of every blue box in the warehouse. If you want your specific blue box, you must pick it out of the bucket yourself.

An activerecord relation returned by the where method is like the bucket mentioned above.