How do I write a basic update method?
the Activerecord update method is like a guy in a warehouse who's
job it is to take things out of boxes and replace them with new contents. You give him a list of identification numbers, and the contents of those boxes that are supposed to be replaced.  He takes that list, finds the box, and throws out the stuff he is supposed to throw out and puts in the stuff he is supposed to put in.