Why do you use the Activerecord where method?
Where is for when you want to get a bunch of

Activerecord where is like picking all of the red skittles out of a bag of skittles.  Each skittle is like a single row in a table in the database. the bag of skittles is like a table in the database.  A red skittle is like a particular kind of record in the database.  The act of picking out all of the red skittles is like using the where command to find a particular kind of record in a table.