How do I update a single attribute?
Use the attribute accessor.

Analogy: using the attribute accessor to update an attribute is like mopping the floor and forgetting to tell anyone that you did it.  You made a change, but since nobody else knows about it, it's very likely that nobody will notice and someone else will come along and mop the floor again.  The act of telling somebody that you mopped the floor is like saving the updated attribute to the database.

Note: this is not actually saved to the database.