FAILING TEST: Why is it important to write a user story before you code?
An ounce of prevention is worth a pound of cure.

A user story is like deciding you are going to spent $3 to try a new tea before spending $100 buying a year's supply of tea and discovering that you hate the taste.

This test is a perfect example of why a user story is helpful.  The genereal flow for an application is as follows:

1. Write a user story
2. Draw wireframes.
3. Draw a schema
4. Start coding.

I lept directly to "start coding" because I 'knew' that "this association is just a join table, so I'll just plug users on one side and posts on the other and stick comments in the middle."

Since I didn't take the time to design it, my hacker news clone has a bug. Specifically, I can't tell hacker news that I wrote a post.  The post has a space for me to say I wrote it, but this hacker news clone is like a really dumb book editor who insists on editing a book for me, but for some reason thinks that it's not at all important that she make sure my name is on the book, so the book goes to the printer with 'by anonymous' in giant gold letters on the cover.

In this test, I tried to ask the the book who it was written by, and the book shrugged its shoulders and said "durp?"