How do I update a record using update_attribute?
Analogy: the update attribute method is like a guy who works in a warehouse who can be interrupted and will do something as long as you are only asking him to do 'just one small thing.'  A record is a shelf in the warehouse, and an attribute is a box on the shelf.  If you tell the warehouse guy to put one new thing in that one box on that one shelf, he can and will do it!