How do I change all of my specs to accomidate a change in migrations?
Analogy: making a change in the database is like a cold front breaking all of the pipes in a sewer system.  All of my tests broke at once.

I mass fixed them with the subl find and replace command.  It is

   cmd-shift f

In other words, it is annoyingly difficult to remember.