How do I rename an activerecord one-to-many relationship?
Kori's explanation: in order to set up relationshops, activerecord relies on two conventions:
that the foriegn key is named user_id, and that the model, i.e. the classname, is named Users. If you decide to change the foriegn key name, you have to tell the User what foreign key to look for and Post what classname to look for.

Analogy: Activerecord associations are like your locker.  If you decide to put your backpack in a place that is not your locker, you will have to remind yourself where you put your backpack, and you have to write your name on the backpack so that everything knows that the backback belongs to you.