How do I have a complex association with the least amount of renaming?
This association was tricky because there were two things that User.posts
could mean.  User.posts could mean the posts a user has written, or a post
the user had commented on.  As Ryan said as he helped me fix this bug, one
or the other needs to be renamed. I chose to rename the one to many
relationship because it involved changing only one test instead of nearly
a dozen tests.

Analogy: renaming an activerecord association is like creating a nickname for something
to make it easier to remember. For example, it is much easier to remember that a user
has written many posts instead of a user having many posts, so I renamed it to user.writings
instead of user.posts.