How do I go to a post from the user writings page?
Writings flow
On the user writings page, click on the writings link.  You will be taken to the post page, which contains all of the commentary.

Correction: if you have a '/' at teh beginning of your route, it puts the rest of that route at the root of your domain in the url.  If you don't have that '/' at the beginning of your route, it tacks the rest of the address onto the url you currently are on.

Analogy: Having a '/' at the beginning of your route is like giving someone a full address to get to your place because you know they need it.  So you would say 55 west Kentucky street, Chicago, Illinois, USA, Planet Earth.

An absolute route is like having a '/' at the beginning of your route.

Having no '/' at the begining of your route is like giving someone just the number of your house because you know they are almost there.  So you would say, "I'm house number 55" because you know they are on your street.

Tested two other flows:

How do I go to the author page from the user  writings page?
  On the user writings page, click on the source link.
  you will be taken to the original article which is somewhere out there on the interwebs.

How do I go to the source page from the user writings page?
  On the user writings page, click on the tile.  You will be taken to the post page which contains all of the comments for that post.