How do I get to a post from the user comments page?
User comments flow:  On the user comments page, click on the comment which takes you to the post on which the comment was written.

also tested the click on the username flow.

How do I get to a user profile from the user comments page?

On the user comments page, Click on the username with takes you to the user profile page.