How do I edit a comment?
Flow:
On any page with a comment,
  if the comment was written by a currently logged-in user
  add an edit button
When you click the edit button,
  you are taken to the eit comment page.

Edit the comment and click the button.
You are taken back to the page you originally came from.

NOTE: I only implemented this for the Post page, and it takes you
back to the post page. I deliberately ingnored the comments page because
I intend to extract the comment functionality out into what's called a
'partial.'  I haven't implemented a partial before. I'm excited!