How do I start up this application?
A project readme is like a pilot's preflight checklist.  If you don't have one, it is very likely that the project won't even get off the ground.

Point of order: I hate this readme. I really want to go and make it usable for every person that comes behind me. However, I am operating under time pressure and the real goal of this repo is to make sure I pass the exam tomorrow. I resent not feeling like I am allowed to take the time to refactor my documentation so that anybody can learn from this project.