What do I need to know to build an app like this?
Analogy: this list of features is what I needed to learn in order to get the app to it's current state.

Actually, it's a list of technical concepts I needed to learn in order to get the app to its current state.  I left out about the need to grow enough of a spine to tell people to stop overloading me.