How do I add anchors to a github markdown readme?
Github flavored markdown is neat!  It adds an html anchor link to every header.

For example, if I type #outline for crudding

Github will add a link like this:

https://github.com/tararoys/HackerNewsCRUDexample#outline-for-crudding

So all I have to do is send someone that link, and they will get skipped directly to my outline for crudding section in the readme.

This is called a 'relative link.'

A relative link is like me telling a person to go to my address, and then 'go into the blue room.'  the words go into the blue room only make sense if they are already in my house, just like a relative link only takes you somewhere if you are already on the right page.